add mysqlmgr

+++ Project/Kernel/Interface/ITimerMgr.h
@@ -9,7 +9,7 @@ class ITimerMgr    : public core::ICore
 public:
 
     virtual void Process(s32 tick) = 0;
-    virtual void StartTimer( core::ITimer *timer, s32 delay, s32 count, s32 interval, const char *trace) = 0;
+    virtual void StartTimer( core::ITimer *timer, s64 delay, s32 count, s64 interval, const char *trace) = 0;
     virtual void KillTimer( core::ITimer *timer) = 0;
 };
 #endif
