+++ Project/Kernel/Timer/TimerMgr.h
@@ -14,7 +14,7 @@ public:
 	virtual bool Destroy() { return true; };
 
     virtual void Process(s32 tick);
-    virtual void StartTimer(core::ITimer *timer, s32 delay, s32 count, s32 interval, const char *trace);
+    virtual void StartTimer(core::ITimer *timer, s64 delay, s32 count, s64 interval, const char *trace);
     virtual void KillTimer(core::ITimer *timer);
 protected:
 private:
